+++ tests/supports/messages.test.php
@@ -59,7 +59,7 @@ class MessagesTest extends \PHPUnit_Framework_TestCase {
 		$this->assertContains('bye', $serialize);
 		$this->assertContains('Goodbye', $serialize);
 
-		$messages->flash();
+		$messages->store();
 
 		$this->assertTrue(\Session::has('message'));